test-requirements.txt
# Lint and unit test requirements
flake8
os-testr>=0.4.1
charms.reactive
mock>=1.2
coverage>=3.6
requests
git+https://github.com/openstack/charms.openstack.git#egg=charms-openstack
